Split rocket node tests to match new contract structure
